rename setenv_..._wrong
